+++ projects/ngx-material/src/lib/ngx-snackbar/index.ts
@@ -1,2 +1,5 @@
 
-export * from './services/snackbar.service';
+export * from './snackbar-outlet.module';
+export * from './directives';
+export * from './models';
+export * from './services';