+++ projects/ngx-material/src/lib/ngx-snackbar/models/index.ts
@@ -1,0 +1,6 @@
+
+export * from './info-snackbar';
+export * from './snackbar-options';
+export * from './snackbar-context';
+export * from './snackbar-instance';
+export * from './snackbar-silo';