+++ projects/ngx-material/src/lib/ngx-snackbar/models/snackbar-options.ts
@@ -4,4 +4,6 @@ export interface SnackbarOptions<T> {
   styles: string[];
   dismissable: boolean;
   duration: number|undefined;
+  swipeable: boolean;
+  showTimer: boolean;
 }