+++ projects/ngx-material/src/lib/ngx-snackbar/stories/snackbar-preview/snackbar-preview.component.ts
@@ -1,12 +1,12 @@
 import {Component, inject} from '@angular/core';
 import {SnackbarService} from "../../services/snackbar.service";
 import {MatButtonModule} from "@angular/material/button";
-import {SnackbarOutletModule} from "../../snackbar-outlet.module";
+import {NgxSnackbarOutletModule} from "../../snackbar-outlet.module";
 
 @Component({
   selector: 'ngx-snackbar-preview',
   standalone: true,
-  imports: [MatButtonModule, SnackbarOutletModule],
+  imports: [MatButtonModule, NgxSnackbarOutletModule],
   templateUrl: './snackbar-preview.component.html',
   styleUrls: ['./snackbar-preview.component.css']
 })