+++ projects/ngx-tools/src/animations/index.ts
@@ -5,4 +5,5 @@ export * from './hoz-collapse.animation';
 export * from './overlay.animation';
 export * from './move-up.animation';
 export * from './side-menu.animation';
+export * from './snackbar.animation';
 export * from './easings';